commit added/changes

+++ frontend/src/store/useAgentStore.js
@@ -1,7 +1,8 @@
 import { create } from "zustand";
 import { api } from "../api/axios";
+import { toast } from "react-toastify";
 
-export const useAgentStore = create((set) => ({
+export const useAgentStore = create((set, get) => ({
   agents: [],
   isLoading: false,
   error: null,
@@ -15,12 +16,14 @@ export const useAgentStore = create((set) => ({
       set({ error, isLoading: false });
     }
   },
+
   createAgent: async (agent) => {
     try {
       set({ isLoading: true });
       const response = await api.post("/agents/create-agent", agent);
-      set({ agents: [...agents, response.data], isLoading: false });
-      toast.success("Created succesful");
+      const currentAgents = get().agents; // Get the current agents from state
+      set({ agents: [...currentAgents, response.data], isLoading: false });
+      toast.success("Created successfully");
     } catch (error) {
       set({ error, isLoading: false });
       toast.error(error.message);
@@ -31,13 +34,14 @@ export const useAgentStore = create((set) => ({
     try {
       set({ isLoading: true });
       const response = await api.put(`/agents/update-agent/${id}`, agent);
+      const currentAgents = get().agents; // Get the current agents from state
       set({
-        agents: agents.map((agent) =>
-          agent.id === id ? response.data : agent,
+        agents: currentAgents.map((existingAgent) =>
+          existingAgent._id === id ? response.data : existingAgent,
         ),
         isLoading: false,
       });
-      toast.success("Updated succesful");
+      toast.success("Updated successfully");
     } catch (error) {
       set({ error, isLoading: false });
       toast.error(error.message);
@@ -47,15 +51,21 @@ export const useAgentStore = create((set) => ({
   deleteAgent: async (id) => {
     try {
       set({ isLoading: true });
-      await api.delete(`/agents/delete-agent/${id}`);
-      set({
-        agents: agents.filter((agent) => agent.id !== id),
-        isLoading: false,
-      });
-      toast.success("Deleted succesful");
+      const currentAgents = get().agents; // Get the current agents from state
+      const agentToDelete = currentAgents.find((agent) => agent._id === id);
+      if (agentToDelete) {
+        await api.delete(`/agents/delete-agent/${id}`);
+        set({
+          agents: currentAgents.filter((agent) => agent._id !== id),
+          isLoading: false,
+        });
+        toast.success(`Agent ${agentToDelete.name} deleted successfully`);
+      } else {
+        throw new Error("Agent not found");
+      }
     } catch (error) {
       set({ error, isLoading: false });
-      toast.error(error.message);
+      toast.error(error.message || "Error deleting agent");
     }
   },
 }));
